requirements.txt
pexpect~=4.8.0
pyyaml~=6.0
pycairo
PyGObject~=3.42.2
IP2Location~=8.9.0
requests~=2.28.2